feat: implementación de carga de datos con switch pagos

+++ switch/src/main/java/ec/edu/espe/banquito/switchpagos/repository/PaymentDetailRepository.java
@@ -1,14 +1,15 @@
 package ec.edu.espe.banquito.switchpagos.repository;
 
-import ec.edu.espe.banquito.switchpagos.model.PaymentDetail;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import ec.edu.espe.banquito.switchpagos.model.PaymentDetail;
 
 @Repository
 public interface PaymentDetailRepository extends JpaRepository<PaymentDetail, Integer> {
 
     // Para procesar el lote iterando "línea por línea" ordenado por el número de línea
-    List<PaymentDetail> findByPaymentBatchIdOrderByLineNumberAsc(Long paymentBatchId);
+    List<PaymentDetail> findByPaymentBatchIdOrderByLineNumberAsc(Integer paymentBatchId);
 }
